added randomstate & initialized regression things (model, train_test_split, OneHotEncoder()    ) to None
not sure why but if i ran fit again, or maybe it was something else after running it for the first time, coefficients spiked to 10^12.
not sure if it was that or the dataframe output which forced sparse (??) ness which fucks up the model?

anyways i put a random state in and initialized to None, and it appears to run repeatably! and there's a correlation! so that's good news